requirements.txt
discord.py
boto3
paramiko
asyncio
aiohttp
python-dotenv
rcon